Inline Case files for FunctionOnlyReturningConstantSpec
